+++ waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/mapping/service/impl/PrefixBasedDatabaseMappingService.java
@@ -31,7 +31,6 @@ import java.util.function.BiFunction;
 import javax.validation.constraints.NotNull;
 
 import org.apache.commons.io.IOUtils;
-import org.apache.hadoop.hive.metastore.api.GetAllFunctionsResponse;
 import org.apache.hadoop.hive.metastore.api.TableMeta;
 import org.apache.logging.log4j.util.Strings;
 import org.slf4j.Logger;
@@ -300,12 +299,6 @@ public class PrefixBasedDatabaseMappingService implements MappingEventListener {
         }
         List<String> result = executeRequests(allRequests, GET_DATABASES_TIMEOUT, "Can't fetch databases: {}");
         return result;
-      }
-
-      @Override
-      public GetAllFunctionsResponse getAllFunctions() {
-        List<DatabaseMapping> databaseMappings = getDatabaseMappings();
-        return super.getAllFunctions(databaseMappings);
       }
     };
   }