+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## TBD
+### Fixed
+* Fixed rest api (http://localhost:18000/api/admin/federations/) which broke in 3.3.0 release.
+
 ## [3.3.0] - 2019-04-30
 ### Fixed
 * Reconnection to metastores for MANUAL database resolution. With this change, the server configuration can specify a `status-polling-delay` and `status-polling-delay-time-unit`. See the [README](https://github.com/HotelsDotCom/waggle-dance#server) for more information.