+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## TBD
 ### Fixed
+* Fixed REST API (http://localhost:18000/api/admin/federations/) which broke in 3.3.0 release.
 * Prefixing of UDF's used in a view. See [#165](https://github.com/HotelsDotCom/waggle-dance/issues/165).
 
 ## [3.3.0] - 2019-04-30