+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## TBD
 ### Fixed
-* Show Functions now shows UDF's from all metastores. See [#164](https://github.com/HotelsDotCom/waggle-dance/issues/164).
+* `Show Functions` now shows UDFs from all metastores. See [#164](https://github.com/HotelsDotCom/waggle-dance/issues/164).
 * Fixed REST API (http://localhost:18000/api/admin/federations/) which broke in 3.3.0 release.
-* Prefixing of UDF's used in a view. See [#165](https://github.com/HotelsDotCom/waggle-dance/issues/165).
+* Prefixing of UDFs used in a view. See [#165](https://github.com/HotelsDotCom/waggle-dance/issues/165).
 
 ## [3.3.0] - 2019-04-30
 ### Fixed